refactor whole structure, instead of inheritance make it version-separative
