+++ powerfactory_tools/versions/pf2022/__init__.py
@@ -1,9 +1,9 @@
 # :author: Sasan Jacob Rasti <sasan_jacob.rasti@tu-dresden.de>
-# :copyright: Copyright (c) Institute of Electrical Power Systems and High Voltage Engineering - TU Dresden, 2022-2023.
+# :copyright: Copyright (c) Institute of Electrical Power Systems and High Voltage Engineering - TU Dresden, 2022-2024.
 # :license: BSD 3-Clause
 
-from powerfactory_tools.versions.pf2022.exporter import PowerFactoryExporter
-from powerfactory_tools.versions.pf2022.exporter import export_powerfactory_data
+from powerfactory_tools.versions.pf2022.exporter.exporter import PowerFactoryExporter
+from powerfactory_tools.versions.pf2022.exporter.exporter import export_powerfactory_data
 from powerfactory_tools.versions.pf2022.interface import PowerFactoryInterface
 
 __all__ = [
