+++ powerfactory_tools/versions/pf2022/constants.py
@@ -1,14 +1,14 @@
 # :author: Sasan Jacob Rasti <sasan_jacob.rasti@tu-dresden.de>
 # :author: Sebastian Krahmer <sebastian.krahmer@tu-dresden.de>
-# :copyright: Copyright (c) Institute of Electrical Power Systems and High Voltage Engineering - TU Dresden, 2022-2023.
+# :copyright: Copyright (c) Institute of Electrical Power Systems and High Voltage Engineering - TU Dresden, 2022-2024.
 # :license: BSD 3-Clause
 
 from __future__ import annotations
 
 import typing as t
 
-from powerfactory_tools.base.types import Currency
-from powerfactory_tools.base.types import MetricPrefix
+from powerfactory_tools.versions.pf2022.types import Currency
+from powerfactory_tools.versions.pf2022.types import MetricPrefix
 
 if t.TYPE_CHECKING:
     from collections.abc import Sequence
